Testing: Add a custom test runner for unit tests
This makes it possible to run unit tests on the esp32s3
using `cargo test`.

Instead of asserting failures, tests are modified to return
an anyhow::Error. This needs to be done because on the esp32s3,
failed assertions cause an immediate abort panic, which prevents
other tests from being run.

This problem is not solved completely by this approach, tests can
of course still panic by accident.

I also looked into defmt-test instead of messing with the test runner,
but it seems that they don't support unit tests, so with just this
approach testing private code would be impossible.