Config: Store RX/TX Whitelist instead of single blacklist
Previously it was not really clear from the blacklist only,
which nodes were allowed to receive/transmit from other nodes.

The blacklist is now instead split into two whitelists:

1. rx_whitelist: lists which nodes may overhear packets from
   other nodes.
2. tx_whitelist: lists which nodes may send/relay packets to
   other nodes.

I used whitelists instead of blacklist, since the TrafficGenerators
need to know which receivers are valid, which is easier with a
whitelist. Functionally there is no difference.