ESP: Prevent (possible) compiler optimization related to received_packets
The receive callback actually works, but calls to receive always
return None. It could be that the compiler optimizes out the calls,
because the reception callback is always executed asynchronously
(caused by an interrupt?).

To prevent this, received_packets.pop_front() is enclosed by a black_box(),
which prevents the compiler from optimizing anything inside.