EspChannel: Convert WiFi frames correctly
The header is compactly stored as a bitfield, when received.
The struct documentation for Espressif WiFi only contains documentation
for the already decoded header.

This is the reason for the crashes, since my binary format was just
wrong because of this. I got the correct bitfield accesses from the
Rust bindings of wifi_pkt_rx_ctrl_t, so hopefully everything should
work now.