ESP: Remove all unwrap()'s from user code
This makes sure that we will not panic because of a mistake in our code.
I have left unwrap()'s on locks and clones, since we can't do anything
better about those.

Of course, not all parts of the code now have correct error handling. I
have added TODOs everywhere things still need to be improved.