Bugfix: Off-by-One in Frame encoding
Previously, the frame count was appended to the end of the header. This
does not make sense, because the header is preallocated at the start to
be exactly HEADER size bytes long. So we were writing header data beyond
the header, which can of course not be correctly decoded on the other
side.

Again, some simple encoding/decoding tests would have caught this bug early.