Stats: Move statistic tracking into Channel and enable logging on ESP
By moving the statistic tracking into the Channels, we do not have to
manually call add_rec or add_send anymore.

For logging on the ESP, we provide a new StatsLogger trait. In the
simulator, the behaviour is the same as previously i.e. logging to a
file. On the ESP, we print "STATS filename data" to stdout, so a
bash/python script running on the host can easily filter out statistics
logs and write them to a file as well.

The script for this still needs to be written, but this will be part of
another commit.