Stats: Rework data layout again
We now store a single NodeID instead of a seperate column for every
possible nexthop, since we only need to consider a single node for each
datapoint added and can just change the NodeID between datapoints.

We now also log our own NodeID, the Traffic Generator and granular
information related to package transmission/reception as well as cache
information.

This commit broke the rx_tx_barchart plot and I have not figured out how
to fix that yet, so that will have to be done in another commit.