ESP: Store packet hash as magic instead of random value
Previously, storing partial frames in case of transmission failure did
no make any sense, as the magic value would be a different one on
retransmission. So the retransmitted frames could not be used to
complete the original partial frames.

We now save a 32bit packet hash as the magic value, which is constant
across retransmissions and should solve this problem.