TrafficGenerator: Add debug prints for checking achieved throughput
The poisson and random generators currently set the next generation
timestamp in the future based on the targeted throughput. But this will
only work correctly, if the generator is called at exactly that point in
the time.

This commit adds debug prints to both generators to check if this
happens in practice.