TrafficGenerator: Add custom packet size distribution
The distribution is bimodal with peaks at 60 and 1400 Bytes. These
values were determined using the paper "Quantifying the deployment
of TCP options - a comparative study", which looked at packet size
distributions of TCP traffic.

These values may not be 100% accurate for our application, but it is
currently more important to generate larger packets at all. The
distribution itself can be improved later.