EspChannel: Use BTreeMap to store FrameCollections instead of HashMap
A lot of the crashes I have seen with high traffic generation happen
when trying to allocate memory for the FrameCollection backing store.

The problem is, that HashMaps have too much overhead.
Rusts HashMap internally has a load factor property, which causes the
HashMap to automatically resize if this value is reached. Sometimes it
needed to allocate an extra continuous chunk of ~10kB, for which the
allocator could not find a free slot.

I found a [post](https://ntietz.com/blog/rust-hashmap-overhead/) showing
that a HashMap can have up to 2x overhead, compared to 0.5x overhead of a
BTreeMap.

As we do not store a lot of FrameCollections, the performance overhead
of using a BTreeMap should be negligible. More importantly, it does not
store elements in one continuous array, which should lead to smaller
reallocations.