FrameCollection: Allow out-of-order insertion of first frame
Previously, adding following frames to a FrameCollection before the
first frame failed, because the amount of total frames inside the
FrameCollection is then set to zero.

So we implicitly assumed that the first frame will always be added to
the collection first. This is not always true, for example when the
first frame is lost and has to be retransmitted.

This commit adds additional checks to catch and handle this case, so
frames can now be added in any order. I also added some additional tests
for this case specifically.